+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### next
 - add support for backtab key (by default it's bound to :previous_match)
+- `:rename` built-in verb, best used with its keyboard shortcut F2
+- new verb argument: `auto_exec`
+- new standard verb arguments: `{file-stem}`, `{file-extension}`, and `{file-dot-extension}`,
 - fix a few minor bugs
 
 <a name="v1.4.0"></a>